Restructure the code in util folders where necceserry
