+++ backend/src/util/flatbuffer/flatbuffer_message_builder.h
@@ -1,7 +1,7 @@
 #ifndef FLATBUFFER_MESSAGE_BUILDER_H
 #define FLATBUFFER_MESSAGE_BUILDER_H
 
-#include "flatbuffer/messageClass_generated.h"
+#include "../../flatbuffer/messageClass_generated.h"
 
 namespace FlatbufferMessageBuilder {
 std::string buildHostMessage(int room_id);
