+++ backend/src/util/globals.h
@@ -1,8 +1,8 @@
 #ifndef GLOBALS_H
 #define GLOBALS_H
 
-#include "client_repository.h"
-#include "party_repository.h"
+#include "../client_repository.h"
+#include "../party_repository.h"
 
 inline PartyRepository party_repository;
 inline ClientRepository client_repository;
